Dropdown list displays Gateway languages
