+++ package.json
@@ -56,6 +56,7 @@
     "electron": "1.7.9",
     "electron-packager": "9.1.0",
     "electron-prebuilt-compile": "1.6.11",
+    "electron-react-devtools": "^0.5.3",
     "enzyme": "2.9.1",
     "eslint": "2.13.1",
     "eslint-plugin-jest": "21.3.2",
